requirements.txt
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
dateutils==0.6.12
decorator==5.0.9
Django==3.2.7
django-extensions==3.1.3
djangorestframework==3.12.4
docopt==0.6.2
fancycompleter==0.9.1
freezegun==1.1.0
frogress==0.9.1
idna==3.2
iniconfig==1.1.1
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.3
packaging==21.0
parso==0.8.2
pdbpp==0.10.3
petl==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
py==1.10.0
Pygments==2.10.0
pyparsing==2.4.7
pyrepl==0.9.0
pytest==6.2.5
pytest-django==4.4.0
pytest-watch==4.2.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.5.0
six==1.8.0
sqlparse==0.4.1
swapi==0.1.3
toml==0.10.2
traitlets==5.1.0
ujson==1.33
urllib3==1.26.6
watchdog==2.1.5
wcwidth==0.2.5
wmctrl==0.4
